Move common classes from android module to pure kotlin module.
