+++ android/src/androidTest/java/tw/ktrssreader/parser/GoogleParserTest.kt
@@ -5,7 +5,7 @@ import org.junit.experimental.runners.Enclosed
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.xmlpull.v1.XmlPullParserException
-import tw.ktrssreader.model.channel.GoogleChannelData
+import tw.ktrssreader.kotlin.model.channel.GoogleChannelData
 import tw.ktrssreader.parser.ChannelItemTestData.FULL_GOOGLE_CHANNEL
 import tw.ktrssreader.parser.ChannelItemTestData.GOOGLE_PLAY_FOLDER
 import tw.ktrssreader.parser.ChannelItemTestData.PARTIAL_GOOGLE_CHANNEL
