+++ android/src/main/java/tw/ktrssreader/parser/AutoMixParser.kt
@@ -18,69 +18,69 @@ package tw.ktrssreader.parser
 
 import org.xmlpull.v1.XmlPullParser
 import org.xmlpull.v1.XmlPullParserException
-import tw.ktrssreader.constant.ParserConst
-import tw.ktrssreader.constant.ParserConst.AUTHOR
-import tw.ktrssreader.constant.ParserConst.BLOCK
-import tw.ktrssreader.constant.ParserConst.CATEGORY
-import tw.ktrssreader.constant.ParserConst.CLOUD
-import tw.ktrssreader.constant.ParserConst.COMMENTS
-import tw.ktrssreader.constant.ParserConst.COPYRIGHT
-import tw.ktrssreader.constant.ParserConst.DESCRIPTION
-import tw.ktrssreader.constant.ParserConst.DOCS
-import tw.ktrssreader.constant.ParserConst.DOMAIN
-import tw.ktrssreader.constant.ParserConst.EMAIL
-import tw.ktrssreader.constant.ParserConst.ENCLOSURE
-import tw.ktrssreader.constant.ParserConst.EXPLICIT
-import tw.ktrssreader.constant.ParserConst.GENERATOR
-import tw.ktrssreader.constant.ParserConst.GOOGLE_AUTHOR
-import tw.ktrssreader.constant.ParserConst.GOOGLE_BLOCK
-import tw.ktrssreader.constant.ParserConst.GOOGLE_CATEGORY
-import tw.ktrssreader.constant.ParserConst.GOOGLE_DESCRIPTION
-import tw.ktrssreader.constant.ParserConst.GOOGLE_EMAIL
-import tw.ktrssreader.constant.ParserConst.GOOGLE_EXPLICIT
-import tw.ktrssreader.constant.ParserConst.GOOGLE_IMAGE
-import tw.ktrssreader.constant.ParserConst.GOOGLE_OWNER
-import tw.ktrssreader.constant.ParserConst.GUID
-import tw.ktrssreader.constant.ParserConst.HEIGHT
-import tw.ktrssreader.constant.ParserConst.HREF
-import tw.ktrssreader.constant.ParserConst.IMAGE
-import tw.ktrssreader.constant.ParserConst.ITEM
-import tw.ktrssreader.constant.ParserConst.ITUNES_AUTHOR
-import tw.ktrssreader.constant.ParserConst.ITUNES_BLOCK
-import tw.ktrssreader.constant.ParserConst.ITUNES_CATEGORY
-import tw.ktrssreader.constant.ParserConst.ITUNES_COMPLETE
-import tw.ktrssreader.constant.ParserConst.ITUNES_DURATION
-import tw.ktrssreader.constant.ParserConst.ITUNES_EMAIL
-import tw.ktrssreader.constant.ParserConst.ITUNES_EPISODE
-import tw.ktrssreader.constant.ParserConst.ITUNES_EPISODE_TYPE
-import tw.ktrssreader.constant.ParserConst.ITUNES_EXPLICIT
-import tw.ktrssreader.constant.ParserConst.ITUNES_IMAGE
-import tw.ktrssreader.constant.ParserConst.ITUNES_NAME
-import tw.ktrssreader.constant.ParserConst.ITUNES_NEW_FEED_URL
-import tw.ktrssreader.constant.ParserConst.ITUNES_OWNER
-import tw.ktrssreader.constant.ParserConst.ITUNES_SEASON
-import tw.ktrssreader.constant.ParserConst.ITUNES_TITLE
-import tw.ktrssreader.constant.ParserConst.ITUNES_TYPE
-import tw.ktrssreader.constant.ParserConst.LANGUAGE
-import tw.ktrssreader.constant.ParserConst.LAST_BUILD_DATE
-import tw.ktrssreader.constant.ParserConst.LINK
-import tw.ktrssreader.constant.ParserConst.MANAGING_EDITOR
-import tw.ktrssreader.constant.ParserConst.OWNER
-import tw.ktrssreader.constant.ParserConst.PUB_DATE
-import tw.ktrssreader.constant.ParserConst.RATING
-import tw.ktrssreader.constant.ParserConst.SKIP_DAYS
-import tw.ktrssreader.constant.ParserConst.SKIP_HOURS
-import tw.ktrssreader.constant.ParserConst.SOURCE
-import tw.ktrssreader.constant.ParserConst.TEXT
-import tw.ktrssreader.constant.ParserConst.TEXT_INPUT
-import tw.ktrssreader.constant.ParserConst.TITLE
-import tw.ktrssreader.constant.ParserConst.TTL
-import tw.ktrssreader.constant.ParserConst.TYPE
-import tw.ktrssreader.constant.ParserConst.URL
-import tw.ktrssreader.constant.ParserConst.WEB_MASTER
-import tw.ktrssreader.constant.ParserConst.WIDTH
-import tw.ktrssreader.model.channel.*
-import tw.ktrssreader.model.item.*
+import tw.ktrssreader.kotlin.constant.ParserConst
+import tw.ktrssreader.kotlin.constant.ParserConst.AUTHOR
+import tw.ktrssreader.kotlin.constant.ParserConst.BLOCK
+import tw.ktrssreader.kotlin.constant.ParserConst.CATEGORY
+import tw.ktrssreader.kotlin.constant.ParserConst.CLOUD
+import tw.ktrssreader.kotlin.constant.ParserConst.COMMENTS
+import tw.ktrssreader.kotlin.constant.ParserConst.COPYRIGHT
+import tw.ktrssreader.kotlin.constant.ParserConst.DESCRIPTION
+import tw.ktrssreader.kotlin.constant.ParserConst.DOCS
+import tw.ktrssreader.kotlin.constant.ParserConst.DOMAIN
+import tw.ktrssreader.kotlin.constant.ParserConst.EMAIL
+import tw.ktrssreader.kotlin.constant.ParserConst.ENCLOSURE
+import tw.ktrssreader.kotlin.constant.ParserConst.EXPLICIT
+import tw.ktrssreader.kotlin.constant.ParserConst.GENERATOR
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_AUTHOR
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_BLOCK
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_CATEGORY
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_DESCRIPTION
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_EMAIL
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_EXPLICIT
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_IMAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.GOOGLE_OWNER
+import tw.ktrssreader.kotlin.constant.ParserConst.GUID
+import tw.ktrssreader.kotlin.constant.ParserConst.HEIGHT
+import tw.ktrssreader.kotlin.constant.ParserConst.HREF
+import tw.ktrssreader.kotlin.constant.ParserConst.IMAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITEM
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_AUTHOR
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_BLOCK
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_CATEGORY
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_COMPLETE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_DURATION
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_EMAIL
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_EPISODE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_EPISODE_TYPE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_EXPLICIT
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_IMAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_NAME
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_NEW_FEED_URL
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_OWNER
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_SEASON
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_TITLE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITUNES_TYPE
+import tw.ktrssreader.kotlin.constant.ParserConst.LANGUAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.LAST_BUILD_DATE
+import tw.ktrssreader.kotlin.constant.ParserConst.LINK
+import tw.ktrssreader.kotlin.constant.ParserConst.MANAGING_EDITOR
+import tw.ktrssreader.kotlin.constant.ParserConst.OWNER
+import tw.ktrssreader.kotlin.constant.ParserConst.PUB_DATE
+import tw.ktrssreader.kotlin.constant.ParserConst.RATING
+import tw.ktrssreader.kotlin.constant.ParserConst.SKIP_DAYS
+import tw.ktrssreader.kotlin.constant.ParserConst.SKIP_HOURS
+import tw.ktrssreader.kotlin.constant.ParserConst.SOURCE
+import tw.ktrssreader.kotlin.constant.ParserConst.TEXT
+import tw.ktrssreader.kotlin.constant.ParserConst.TEXT_INPUT
+import tw.ktrssreader.kotlin.constant.ParserConst.TITLE
+import tw.ktrssreader.kotlin.constant.ParserConst.TTL
+import tw.ktrssreader.kotlin.constant.ParserConst.TYPE
+import tw.ktrssreader.kotlin.constant.ParserConst.URL
+import tw.ktrssreader.kotlin.constant.ParserConst.WEB_MASTER
+import tw.ktrssreader.kotlin.constant.ParserConst.WIDTH
+import tw.ktrssreader.kotlin.model.channel.*
+import tw.ktrssreader.kotlin.model.item.*
 import tw.ktrssreader.utils.logD
 import java.io.IOException
