+++ android/src/main/java/tw/ktrssreader/parser/ParserBase.kt
@@ -19,54 +19,54 @@ package tw.ktrssreader.parser
 import android.util.Xml
 import org.xmlpull.v1.XmlPullParser
 import org.xmlpull.v1.XmlPullParserException
-import tw.ktrssreader.constant.ParserConst.AUTHOR
-import tw.ktrssreader.constant.ParserConst.CATEGORY
-import tw.ktrssreader.constant.ParserConst.CHANNEL
-import tw.ktrssreader.constant.ParserConst.CLOUD
-import tw.ktrssreader.constant.ParserConst.COMMENTS
-import tw.ktrssreader.constant.ParserConst.COPYRIGHT
-import tw.ktrssreader.constant.ParserConst.DAY
-import tw.ktrssreader.constant.ParserConst.DESCRIPTION
-import tw.ktrssreader.constant.ParserConst.DOCS
-import tw.ktrssreader.constant.ParserConst.DOMAIN
-import tw.ktrssreader.constant.ParserConst.ENCLOSURE
-import tw.ktrssreader.constant.ParserConst.GENERATOR
-import tw.ktrssreader.constant.ParserConst.GUID
-import tw.ktrssreader.constant.ParserConst.HEIGHT
-import tw.ktrssreader.constant.ParserConst.HOUR
-import tw.ktrssreader.constant.ParserConst.IMAGE
-import tw.ktrssreader.constant.ParserConst.ITEM
-import tw.ktrssreader.constant.ParserConst.LANGUAGE
-import tw.ktrssreader.constant.ParserConst.LAST_BUILD_DATE
-import tw.ktrssreader.constant.ParserConst.LENGTH
-import tw.ktrssreader.constant.ParserConst.LINK
-import tw.ktrssreader.constant.ParserConst.MANAGING_EDITOR
-import tw.ktrssreader.constant.ParserConst.NAME
-import tw.ktrssreader.constant.ParserConst.PATH
-import tw.ktrssreader.constant.ParserConst.PERMALINK
-import tw.ktrssreader.constant.ParserConst.PORT
-import tw.ktrssreader.constant.ParserConst.PROTOCOL
-import tw.ktrssreader.constant.ParserConst.PUB_DATE
-import tw.ktrssreader.constant.ParserConst.RATING
-import tw.ktrssreader.constant.ParserConst.REGISTER_PROCEDURE
-import tw.ktrssreader.constant.ParserConst.SKIP_DAYS
-import tw.ktrssreader.constant.ParserConst.SKIP_HOURS
-import tw.ktrssreader.constant.ParserConst.SOURCE
-import tw.ktrssreader.constant.ParserConst.TEXT_INPUT
-import tw.ktrssreader.constant.ParserConst.TITLE
-import tw.ktrssreader.constant.ParserConst.TTL
-import tw.ktrssreader.constant.ParserConst.TYPE
-import tw.ktrssreader.constant.ParserConst.URL
-import tw.ktrssreader.constant.ParserConst.WEB_MASTER
-import tw.ktrssreader.constant.ParserConst.WIDTH
-import tw.ktrssreader.model.channel.*
-import tw.ktrssreader.model.item.*
+import tw.ktrssreader.kotlin.constant.ParserConst.AUTHOR
+import tw.ktrssreader.kotlin.constant.ParserConst.CATEGORY
+import tw.ktrssreader.kotlin.constant.ParserConst.CHANNEL
+import tw.ktrssreader.kotlin.constant.ParserConst.CLOUD
+import tw.ktrssreader.kotlin.constant.ParserConst.COMMENTS
+import tw.ktrssreader.kotlin.constant.ParserConst.COPYRIGHT
+import tw.ktrssreader.kotlin.constant.ParserConst.DAY
+import tw.ktrssreader.kotlin.constant.ParserConst.DESCRIPTION
+import tw.ktrssreader.kotlin.constant.ParserConst.DOCS
+import tw.ktrssreader.kotlin.constant.ParserConst.DOMAIN
+import tw.ktrssreader.kotlin.constant.ParserConst.ENCLOSURE
+import tw.ktrssreader.kotlin.constant.ParserConst.GENERATOR
+import tw.ktrssreader.kotlin.constant.ParserConst.GUID
+import tw.ktrssreader.kotlin.constant.ParserConst.HEIGHT
+import tw.ktrssreader.kotlin.constant.ParserConst.HOUR
+import tw.ktrssreader.kotlin.constant.ParserConst.IMAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.ITEM
+import tw.ktrssreader.kotlin.constant.ParserConst.LANGUAGE
+import tw.ktrssreader.kotlin.constant.ParserConst.LAST_BUILD_DATE
+import tw.ktrssreader.kotlin.constant.ParserConst.LENGTH
+import tw.ktrssreader.kotlin.constant.ParserConst.LINK
+import tw.ktrssreader.kotlin.constant.ParserConst.MANAGING_EDITOR
+import tw.ktrssreader.kotlin.constant.ParserConst.NAME
+import tw.ktrssreader.kotlin.constant.ParserConst.PATH
+import tw.ktrssreader.kotlin.constant.ParserConst.PERMALINK
+import tw.ktrssreader.kotlin.constant.ParserConst.PORT
+import tw.ktrssreader.kotlin.constant.ParserConst.PROTOCOL
+import tw.ktrssreader.kotlin.constant.ParserConst.PUB_DATE
+import tw.ktrssreader.kotlin.constant.ParserConst.RATING
+import tw.ktrssreader.kotlin.constant.ParserConst.REGISTER_PROCEDURE
+import tw.ktrssreader.kotlin.constant.ParserConst.SKIP_DAYS
+import tw.ktrssreader.kotlin.constant.ParserConst.SKIP_HOURS
+import tw.ktrssreader.kotlin.constant.ParserConst.SOURCE
+import tw.ktrssreader.kotlin.constant.ParserConst.TEXT_INPUT
+import tw.ktrssreader.kotlin.constant.ParserConst.TITLE
+import tw.ktrssreader.kotlin.constant.ParserConst.TTL
+import tw.ktrssreader.kotlin.constant.ParserConst.TYPE
+import tw.ktrssreader.kotlin.constant.ParserConst.URL
+import tw.ktrssreader.kotlin.constant.ParserConst.WEB_MASTER
+import tw.ktrssreader.kotlin.constant.ParserConst.WIDTH
+import tw.ktrssreader.kotlin.model.channel.*
+import tw.ktrssreader.kotlin.model.item.*
 import tw.ktrssreader.utils.logD
 import tw.ktrssreader.utils.logW
 import java.io.ByteArrayInputStream
 import java.io.IOException
 
-abstract class ParserBase<out T : RssStandardChannel> : Parser<T> {
+abstract class ParserBase<out T : RssStandardChannel> : tw.ktrssreader.kotlin.parser.Parser<T> {
 
     abstract val logTag: String
