+++ android/src/main/java/tw/ktrssreader/parser/RssStandardParser.kt
@@ -17,7 +17,7 @@
 package tw.ktrssreader.parser
 
 import org.xmlpull.v1.XmlPullParserException
-import tw.ktrssreader.model.channel.RssStandardChannel
+import tw.ktrssreader.kotlin.model.channel.RssStandardChannel
 
 class RssStandardParser : ParserBase<RssStandardChannel>() {
