+++ app/src/main/java/tw/ktrssreader/reader/strategy/StandardRssStrategy.kt
@@ -2,7 +2,7 @@ package tw.ktrssreader.reader.strategy
 
 import kotlinx.coroutines.flow.Flow
 import tw.ktrssreader.Reader
-import tw.ktrssreader.model.channel.RssStandardChannelData
+import tw.ktrssreader.kotlin.model.channel.RssStandardChannelData
 import java.io.Serializable
 import java.nio.charset.Charset
